Close all files at the end of the document (fixes #12)
I was only closing half of the files (and not quite correctly: it
produced some files named things like 4.tex).  The two testfiles
morewrites005 and morewrites006 test respectively whether files
whose material is are kept in memory until being written, and
files that are assigned to a TeX stream directly, are correctly
closed.